requirements.txt
APScheduler==3.5.3
python-dotenv==0.9.1
fake-useragent==0.1.11
html-sanitizer==1.6.3
Scrapy==1.5.1
scrapy-fake-useragent==1.1.0
google-cloud-firestore==0.30.0
Twisted==18.9.0
pytz==2018.7
pytelegrambotapi==3.6.6
cloudscraper==1.2.14
scrapy_sentry==0.9.0
python-dateutil==2.8.1
-e git+https://github.com/yall/scrapy-twitter.git@77b2c523aefffc4cb27408afda046d3340b15174#egg=scrapy_twitter